+++ pytest_container/__init__.py
@@ -5,6 +5,7 @@ images or software in container images with pytest.
 from .build import GitRepositoryBuild
 from .build import MultiStageBuild
 from .container import Container
+from .container import container_and_marks_from_pytest_param
 from .container import container_from_pytest_param
 from .container import container_to_pytest_param
 from .container import DerivedContainer